requirements.txt
# django
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2

# parser
lxml==4.6.3
requests==2.26.0
beautifulsoup4==4.10.0